refactor: switch to `DuplexBlobProvider`
