+++ packages/store/src/persistence/blob/cloud-sync-manager.ts
@@ -1,122 +1,10 @@
+import { Signal, sleep } from '@blocksuite/global/utils';
 import ky from 'ky';
-import { assertExists, Signal, sleep } from '@blocksuite/global/utils';
-
-import type { BlobId, BlobProvider, BlobURL, IDBInstance } from './types.js';
-import { getDatabase, sha } from './utils.js';
+import type { BlobOptionsGetter } from './duplex-provider.js';
+import type { BlobId, BlobURL, IDBInstance } from './types.js';
+import { getDatabase } from './utils.js';
 
 const RETRY_TIMEOUT = 500;
-
-export type BlobOptions = Record<'api' | 'token', string>;
-
-export type BlobOptionsGetter = (key: keyof BlobOptions) => string | undefined;
-
-interface BlobProviderStatic {
-  init(
-    workspace: string,
-    optionsGetter?: BlobOptionsGetter
-  ): Promise<BlobProvider>;
-}
-
-function staticImplements<T>() {
-  return <U extends T>(constructor: U) => constructor;
-}
-
-@staticImplements<BlobProviderStatic>()
-export class IndexedDBBlobProvider implements BlobProvider {
-  private readonly _database: IDBInstance;
-  private readonly _cloud?: CloudSyncManager;
-  public uploading = false;
-
-  readonly blobs: Set<string> = new Set();
-  readonly signals = {
-    blobAdded: new Signal<BlobId>(),
-    blobDeleted: new Signal<BlobId>(),
-    uploadStateChanged: new Signal<boolean>(),
-    uploadFinished: new Signal<BlobId>(),
-  };
-
-  static async init(
-    workspace: string,
-    optionsGetter?: BlobOptionsGetter
-  ): Promise<IndexedDBBlobProvider> {
-    const provider = new IndexedDBBlobProvider(workspace, optionsGetter);
-    await provider._initBlobs();
-    return provider;
-  }
-
-  private async _initBlobs() {
-    const entries = await this._database.keys();
-    for (const key of entries) {
-      const blobId = key as BlobId;
-      this.signals.blobAdded.emit(blobId);
-      this.blobs.add(blobId);
-    }
-  }
-
-  private constructor(workspace: string, optionsGetter?: BlobOptionsGetter) {
-    this._database = getDatabase('blob', workspace);
-
-    const endpoint = optionsGetter?.('api');
-    if (endpoint) {
-      assertExists(optionsGetter);
-      this.signals.uploadStateChanged.on(uploading => {
-        this.uploading = uploading;
-      });
-      this._cloud = new CloudSyncManager(
-        workspace,
-        optionsGetter,
-        this._database,
-        this.signals.uploadStateChanged,
-        this.signals.uploadFinished
-      );
-    }
-  }
-
-  async get(id: BlobId): Promise<BlobURL | null> {
-    const blob = await this._database.get(id);
-    if (!blob) {
-      const blob = this._cloud?.get(id);
-      if (blob) {
-        this.signals.blobAdded.emit(id);
-        this.blobs.add(id);
-
-        return blob;
-      }
-      return null;
-    }
-
-    const result = URL.createObjectURL(new Blob([blob]));
-    return result;
-  }
-
-  async set(blob: Blob): Promise<BlobId> {
-    const buffer = await blob.arrayBuffer();
-    const hash = await sha(buffer);
-    if (!(await this._database.has(hash))) {
-      await this._database.set(hash, buffer);
-
-      this.signals.blobAdded.emit(hash);
-      this.blobs.add(hash);
-    }
-
-    this._cloud?.addTask(hash, 'add');
-
-    return hash;
-  }
-
-  async delete(id: BlobId): Promise<void> {
-    await this._database.delete(id);
-
-    this.signals.blobDeleted.emit(id);
-    this.blobs.delete(id);
-    this._cloud?.addTask(id, 'delete');
-  }
-
-  async clear(): Promise<void> {
-    await this._database.clear();
-    this.blobs.clear();
-  }
-}
 
 type SyncTask = {
   id: BlobId;
